+++ Assets/Tea Framework/Scripts/Runtime/Input/PlayerInputISystem.cs
@@ -3,6 +3,7 @@ using Unity.Mathematics;
 using UnityEngine;
 
 namespace TeaFramework {
+    
     /// <summary>
     /// System that only RW the values from an input system 
     /// </summary>