+++ Packages/com.datea83.choyutilities/Runtime/Monobehaviour/Generic/GenericParticleManager.cs
@@ -1,79 +1,46 @@
-using System;
-using System.Collections.Generic;
+﻿using System;
 using Unity.Mathematics;
 using UnityEngine;
+using UnityEngine.Pool;
 
 namespace EugeneC.Singleton {
-    public abstract class GenericParticleManager<TEnum, TMono> : GenericSingleton<TMono>
+    
+    public abstract class GenericParticleManager<TEnum, TMono> : GenericPoolingManager<TEnum, ParticleSystem, TMono> 
         where TEnum : Enum
-        where TMono : MonoBehaviour {
-        [Serializable]
-        public struct ParticleSerialize {
-            public TEnum id;
-            public ParticleSystem particle;
-        }
-
-        [Serializable]
-        protected struct SystemSerialize {
-            public ParticleSystem[] particles;
-            public int currentIndex;
-            public int previousIndex;
-        }
-
-        [SerializeField] protected ParticleSerialize[] particleEffects;
-
-        [Tooltip("The pool count meant each particle serialize, meaning total = poolCount * particleEffects.Length")]
-        [SerializeField]
-        protected byte poolCount = 16;
-
-        protected SystemSerialize[] ParticleSystems;
-        protected List<int> PauseIndexes;
-
-        protected virtual async void Start() {
-            try {
-                await Awaitable.WaitForSecondsAsync(.1f, Token);
-
-                ParticleSystems = new SystemSerialize[particleEffects.Length];
-                var currentSystem = 0;
-
-                foreach (var particle in particleEffects) {
-                    ParticleSystems[currentSystem].particles = new ParticleSystem[poolCount];
-                    for (var i = 0; i < poolCount; i++) {
-                        if (particle.particle is null) continue;
-                        var spawn = Instantiate(particle.particle, transform);
-                        ParticleSystems[currentSystem].particles[i] = spawn;
-                    }
-
-                    currentSystem++;
-                }
-            }
-            catch (Exception e) {
-                Debug.LogException(e);
+        where TMono : MonoBehaviour{
+        
+        protected override void Awake() {
+            base.Awake();
+            
+            foreach (var p in poolPrefabs) {
+                var particle = p.prefab;
+                var particleMain = particle.main;
+                particleMain.playOnAwake = false;
             }
         }
 
         public virtual void PlayEffectAtPosition(TEnum id, float3 position) {
-            if (!Enum.IsDefined(typeof(TEnum), id)) return;
-            var index = Array.FindIndex(particleEffects, i => EqualityComparer<TEnum>.Default.Equals(i.id, id));
+            var index = GetPoolIndex(id);
+            if (index == -1) return;
 
-            var particle = ParticleSystems[index].particles[ParticleSystems[index].currentIndex];
+            var particle = RuntimePools[index].spawn[RuntimePools[index].currentIndex];
             particle.transform.position = position;
             particle.Play();
 
-            ParticleSystems[index].previousIndex = ParticleSystems[index].currentIndex;
-            ParticleSystems[index].currentIndex++;
-            ParticleSystems[index].currentIndex %= ParticleSystems[index].particles.Length;
+            RuntimePools[index].previousIndex = RuntimePools[index].currentIndex;
+            RuntimePools[index].currentIndex++;
+            RuntimePools[index].currentIndex %= RuntimePools[index].spawn.Length;
         }
 
         public virtual void PauseAllEffects() {
-            PauseIndexes = new List<int>();
-            for (var i = 0; i < ParticleSystems.Length; i++) {
-                var system = ParticleSystems[i];
-                foreach (var p in system.particles) {
+            PauseIndexes = ListPool<int>.Get();
+            for (var i = 0; i < RuntimePools.Length; i++) {
+                var system = RuntimePools[i];
+                
+                foreach (var p in system.spawn) {
                     if (p.isPlaying)
                         p.Pause();
                 }
-
                 PauseIndexes.Add(i);
             }
         }
@@ -81,14 +48,13 @@ namespace EugeneC.Singleton {
         public virtual void ResumeAllEffects() {
             if (PauseIndexes is null) return;
             foreach (var i in PauseIndexes) {
-                var system = ParticleSystems[i];
-                foreach (var p in system.particles) {
+                var system = RuntimePools[i];
+                foreach (var p in system.spawn) {
                     if (p.isPaused)
                         p.Play();
                 }
             }
-
-            PauseIndexes.Clear();
+            ListPool<int>.Release(PauseIndexes);
         }
     }
 }