+++ Packages/com.datea83.choyutilities/Runtime/Monobehaviour/Generic/GenericSingleton.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using UnityEngine;
 
 namespace EugeneC.Singleton {
+    
     public abstract class GenericSingleton<T> : MonoBehaviour
         where T : MonoBehaviour {
         public static T Instance { get; private set; }