+++ Packages/com.datea83.choyutilities/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.1.4] - 2026-03-15
+
+### Added
+
+* Added GenericPoolingManager
+* Added new GenericAudioManager, GenericParticleManager and GenericUIManager
+* Added GenericSpawnManager
+* HelperCollection, added "RandomValue2" and "RandomValue3"
+
+### Changed
+* Moved legacy GenericAudioManager, GenericParticleManager and GenericUIManager to ObsoleteV2 folder and marked [Obsolete]
+* HelperCollection, all "RandomValue" with GameObject parameter changed to Component instead
+
 ## [0.1.3] - 2026-03-14
 
 ### Changed